refactor: Change column type of email entity
JPA ddl-auto set its type as an enum. When enum value has added in java, JPA does not automatically apply the change to the database that it can lead to operation failure. By explicitly setting its type as a varchar, the change of enum value in java can be done independent to column type.